reorg fixtures (lots of work to do there)
